strip non-slotted typecheck
