+++ src/main.rs
@@ -54,7 +54,7 @@ fn compile(source: &str) -> RuntimeResult<String> {
     use mossy::parser;
     use mossy::preprocessor;
     use mossy::stage::codegen::machine::arch::x86_64;
-    use mossy::stage::{slotted_type_check, type_check, CompilationStage};
+    use mossy::stage::{type_check, CompilationStage};
 
     let input: Vec<(usize, char)> = source.chars().enumerate().collect();
     let pre_processed_input = preprocessor::pre_process(&input)
@@ -62,8 +62,7 @@ fn compile(source: &str) -> RuntimeResult<String> {
 
     parser::parse(&pre_processed_input)
         .map(|program| {
-            //type_check::TypeAnalysis::default()
-            slotted_type_check::TypeAnalysis::default()
+            type_check::TypeAnalysis::default()
                 .and_then(|| x86_64::X86_64)
                 .apply(program)
         })
